add search page and implement PageContext
