+++ templates/core/base2.html
@@ -67,7 +67,7 @@
   </div>
   <div class="container is-flex mt-5">
     <main id="main" class="px-3" style="flex-grow:1;">
-      {% block section %}{% endblock section %}
+      {% block main %}{% endblock main %}
     </main>
 
     <aside role="complementary" class="px-3 is-hidden-mobile ml-5" style="width:25%;flex-shrink:0;">
